+++ docs/changelog.md
@@ -3,6 +3,11 @@
 ## VERSION
 
 * Adds `ProgressChefInfraClient`, `ProgressChefInSpec`, `ProgressChefWorkstation`
+* Adds `ESETSecureAuthentication` [#347](https://github.com/aaronparker/evergreen/issues/347)
+* Updates `WinMerge` to pull updates from GitHub
+* Adds proxy server support for the following functions - `Get-EvergreenApp`, `Save-EvergreenApp`, `Invoke-EvergreenLibraryUpdate`
+
+  Adds `-Proxy` and `-ProxyCredential` arguments. `Get-EvergreenApp` sets an environment variable for the proxy server address and credential for use by internal app and private functions. `Save-EvergreenApp` and `Invoke-EvergreenLibraryUpdate` either use these parameters directly or pass the parameters to other functions
 
 ## 2210.661